requirements.txt
python-dotenv
requests
Jinja2
Pillow
playwright
pytest
playwright==1.44.0
Jinja2==3.1.2
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.0
